+++ NEWS.md
@@ -44,4 +44,4 @@
 
 - Integrate random effects variance parameter out of likelihood.
 - Implement hybrid o-spline and RW model for r(t).
-- Implement logistic model for log r(t).
+- Implement logistic model for log r(t) (rlogistic), and version with random walk (rlogistic_rw).